Add basebot as a class cause @bot.event, etc is annoying
