+++ base/basebot.py
@@ -1,0 +1,170 @@
+"""
+Copyright © Krypton 2019-2023 - https://github.com/kkrypt0nn (https://krypton.ninja)
+Description:
+🐍 A simple template to start to code your own and personalized discord bot in Python programming language.
+
+Version: 5.5.0
+"""
+
+import asyncio
+import os
+import platform
+
+import aiosqlite
+import discord
+from discord.ext import commands
+from discord.ext.commands import Context
+
+import exceptions
+from base.config import CONFIG, Config
+from base.guildedcog import ConfigedBot
+from helpers import BASE_PATH, DATABASE_PATH, SCHEMA_PATH
+from helpers.logger import BotLogger
+
+intents = discord.Intents.all()
+# theoretically redundant but just in case
+intents.typing = False
+intents.dm_typing = False
+intents.guild_typing = False
+intents.message_content = True
+# intents.members = False
+intents.members = True
+intents.presences = True
+intents.guilds = True
+
+
+class BaseBot(ConfigedBot):
+    def __init__(self, config: Config, command_prefix: str, **kwargs):
+        super().__init__(config, command_prefix, **kwargs)
+        self.logger = BotLogger()
+
+    async def on_ready(self) -> None:
+        """
+        The code in this event is executed when the bot is ready.
+        """
+        self.logger.info(f"Logged in as {self.user.name}")
+        self.logger.info(f"discord.py API version: {discord.__version__}")
+        self.logger.info(f"Python version: {platform.python_version()}")
+        self.logger.info(
+            f"Running on: {platform.system()} {platform.release()} ({os.name})"
+        )
+        self.logger.info("-------------------")
+        basic_activity = discord.Activity(
+            name="you", type=discord.ActivityType.watching, status=discord.Status.online
+        )
+        await self.change_presence(activity=basic_activity)
+        # TODO what the heck was this for
+        # if CONFIG.sync_commands_globally:
+        self.logger.info("Syncing commands globally...")
+        # self.tree.clear_commands(guild=self.get_guild(CONFIG.server))
+        commands = await self.tree.sync(guild=self.get_guild(CONFIG.server))
+
+    async def on_message(self, message: discord.Message) -> None:
+        """
+        The code in this event is executed every time someone sends a message, with or without the prefix
+
+        :param message: The message that was sent.
+        """
+        if message.author == self.user or message.author.bot:
+            return
+        await self.process_commands(message)
+
+    async def on_command_completion(self, context: Context) -> None:
+        """
+        The code in this event is executed every time a normal command has been *successfully* executed.
+
+        :param context: The context of the command that has been executed.
+        """
+        full_command_name = context.command.qualified_name
+        split = full_command_name.split(" ")
+        executed_command = str(split[0])
+        if context.guild is not None:
+            self.logger.info(
+                f"Executed {executed_command} command in {context.guild.name} (ID: {context.guild.id}) by {context.author} (ID: {context.author.id})"
+            )
+        else:
+            self.logger.info(
+                f"Executed {executed_command} command by {context.author} (ID: {context.author.id}) in DMs"
+            )
+
+    async def on_command_error(self, context: Context, error) -> None:
+        """
+        The code in this event is executed every time a normal valid command catches an error.
+
+        :param context: The context of the normal command that failed executing.
+        :param error: The error that has been faced.
+        """
+        if isinstance(error, commands.CommandOnCooldown):
+            minutes, seconds = divmod(error.retry_after, 60)
+            hours, minutes = divmod(minutes, 60)
+            hours = hours % 24
+            embed = discord.Embed(
+                description=f"**Please slow down** - You can use this command again in {f'{round(hours)} hours' if round(hours) > 0 else ''} {f'{round(minutes)} minutes' if round(minutes) > 0 else ''} {f'{round(seconds)} seconds' if round(seconds) > 0 else ''}.",
+                color=0xE02B2B,
+            )
+            await context.send(embed=embed)
+        elif isinstance(error, exceptions.UserBlacklisted):
+            """
+            The code here will only execute if the error is an instance of 'UserBlacklisted', which can occur when using
+            the @checks.not_blacklisted() check in your command, or you can raise the error by yourself.
+            """
+            embed = discord.Embed(
+                description="You are blacklisted from using the bot!", color=0xE02B2B
+            )
+            await context.send(embed=embed)
+            self.logger.warning(
+                f"{context.author} (ID: {context.author.id}) tried to execute a command in the guild {context.guild.name} (ID: {context.guild.id}), but the user is blacklisted from using the bot."
+            )
+        elif isinstance(error, exceptions.UserNotOwner):
+            """
+            Same as above, just for the @checks.is_owner() check.
+            """
+            embed = discord.Embed(
+                description="You are not the owner of the bot!", color=0xE02B2B
+            )
+            await context.send(embed=embed)
+            self.logger.warning(
+                f"{context.author} (ID: {context.author.id}) tried to execute an owner only command in the guild {context.guild.name} (ID: {context.guild.id}), but the user is not an owner of the bot."
+            )
+        elif isinstance(error, commands.MissingPermissions):
+            embed = discord.Embed(
+                description="You are missing the permission(s) `"
+                + ", ".join(error.missing_permissions)
+                + "` to execute this command!",
+                color=0xE02B2B,
+            )
+            await context.send(embed=embed)
+        elif isinstance(error, commands.BotMissingPermissions):
+            embed = discord.Embed(
+                description="I am missing the permission(s) `"
+                + ", ".join(error.missing_permissions)
+                + "` to fully perform this command!",
+                color=0xE02B2B,
+            )
+            await context.send(embed=embed)
+        elif isinstance(error, commands.MissingRequiredArgument):
+            embed = discord.Embed(
+                title="Error!",
+                # We need to capitalize because the command arguments have no capital letter in the code.
+                description=str(error).capitalize(),
+                color=0xE02B2B,
+            )
+            await context.send(embed=embed)
+        else:
+            raise error
+
+    async def load_cogs(self) -> None:
+        """
+        The code in this function is executed whenever the bot will start.
+        """
+        for file in (BASE_PATH / "cogs").iterdir():
+            if file.suffix == ".py":
+                extension = file.stem
+                try:
+                    await self.load_extension(f"cogs.{extension}")
+                    self.logger.info(f"Loaded extension '{extension}'")
+                except Exception as e:
+                    exception = f"{type(e).__name__}: {e}"
+                    self.logger.error(
+                        f"Failed to load extension {extension}\n{exception}"
+                    )
