+++ helpers/logger.py
@@ -1,0 +1,58 @@
+import logging
+from logging import Logger
+
+from helpers import LOG_PATH
+
+
+# Setup both of the loggers
+class LoggingFormatter(logging.Formatter):
+    # Colors
+    black = "\x1b[30m"
+    red = "\x1b[31m"
+    green = "\x1b[32m"
+    yellow = "\x1b[33m"
+    blue = "\x1b[34m"
+    gray = "\x1b[38m"
+    # Styles
+    reset = "\x1b[0m"
+    bold = "\x1b[1m"
+
+    COLORS = {
+        logging.DEBUG: gray + bold,
+        logging.INFO: blue + bold,
+        logging.WARNING: yellow + bold,
+        logging.ERROR: red,
+        logging.CRITICAL: red + bold,
+    }
+
+    def format(self, record):
+        log_color = self.COLORS[record.levelno]
+        format = "(black){asctime}(reset) (levelcolor){levelname:<8}(reset) (green){name}(reset) {message}"
+        format = format.replace("(black)", self.black + self.bold)
+        format = format.replace("(reset)", self.reset)
+        format = format.replace("(levelcolor)", log_color)
+        format = format.replace("(green)", self.green + self.bold)
+        formatter = logging.Formatter(format, "%Y-%m-%d %H:%M:%S", style="{")
+        return formatter.format(record)
+
+
+class BotLogger(Logger):
+    def __init__(self):
+        super().__init__("discord_bot", logging.INFO)
+        # Console handler
+        console_handler = logging.StreamHandler()
+        console_handler.setFormatter(LoggingFormatter())
+        # File handler
+        log_path = LOG_PATH / "discord.log"
+        file_handler = logging.FileHandler(log_path, encoding="utf-8", mode="w")
+        file_handler_formatter = logging.Formatter(
+            "[{asctime}] [{levelname:<8}] {name}: {message}",
+            "%Y-%m-%d %H:%M:%S",
+            style="{",
+        )
+        file_handler.setFormatter(file_handler_formatter)
+
+        # Add the handlers
+        self.addHandler(console_handler)
+        self.addHandler(file_handler)
+        self.addHandler(file_handler)
