+++ main.py
@@ -7,19 +7,15 @@ Version: 5.5.0
 """
 
 import asyncio
-import logging
-import os
-import platform
 
 import aiosqlite
 import discord
 from discord.ext import commands
-from discord.ext.commands import Context
 
-import exceptions
+from base.basebot import BaseBot
 from base.config import CONFIG
 from base.guildedcog import ConfigedBot
-from helpers import BASE_PATH, DATABASE_PATH, LOG_PATH, SCHEMA_PATH
+from helpers import DATABASE_PATH, SCHEMA_PATH
 
 intents = discord.Intents.all()
 # theoretically redundant but just in case
@@ -32,64 +28,12 @@ intents.members = True
 intents.presences = True
 intents.guilds = True
 
-bot = ConfigedBot(
+bot = BaseBot(
     CONFIG,
     command_prefix=commands.when_mentioned_or(CONFIG.command_prefix),
     intents=intents,
     help_command=None,
 )
-
-
-# Setup both of the loggers
-class LoggingFormatter(logging.Formatter):
-    # Colors
-    black = "\x1b[30m"
-    red = "\x1b[31m"
-    green = "\x1b[32m"
-    yellow = "\x1b[33m"
-    blue = "\x1b[34m"
-    gray = "\x1b[38m"
-    # Styles
-    reset = "\x1b[0m"
-    bold = "\x1b[1m"
-
-    COLORS = {
-        logging.DEBUG: gray + bold,
-        logging.INFO: blue + bold,
-        logging.WARNING: yellow + bold,
-        logging.ERROR: red,
-        logging.CRITICAL: red + bold,
-    }
-
-    def format(self, record):
-        log_color = self.COLORS[record.levelno]
-        format = "(black){asctime}(reset) (levelcolor){levelname:<8}(reset) (green){name}(reset) {message}"
-        format = format.replace("(black)", self.black + self.bold)
-        format = format.replace("(reset)", self.reset)
-        format = format.replace("(levelcolor)", log_color)
-        format = format.replace("(green)", self.green + self.bold)
-        formatter = logging.Formatter(format, "%Y-%m-%d %H:%M:%S", style="{")
-        return formatter.format(record)
-
-
-logger = logging.getLogger("discord_bot")
-logger.setLevel(logging.INFO)
-
-# Console handler
-console_handler = logging.StreamHandler()
-console_handler.setFormatter(LoggingFormatter())
-# File handler
-log_path = filename = LOG_PATH / "discord.log"
-file_handler = logging.FileHandler(log_path, encoding="utf-8", mode="w")
-file_handler_formatter = logging.Formatter(
-    "[{asctime}] [{levelname:<8}] {name}: {message}", "%Y-%m-%d %H:%M:%S", style="{"
-)
-file_handler.setFormatter(file_handler_formatter)
-
-# Add the handlers
-logger.addHandler(console_handler)
-logger.addHandler(file_handler)
-bot.logger = logger
 
 
 async def init_db():
@@ -99,144 +43,7 @@ async def init_db():
         await db.commit()
 
 
-# TODO can I move all of this somewhere else now?
-@bot.event
-async def on_ready() -> None:
-    """
-    The code in this event is executed when the bot is ready.
-    """
-    bot.logger.info(f"Logged in as {bot.user.name}")
-    bot.logger.info(f"discord.py API version: {discord.__version__}")
-    bot.logger.info(f"Python version: {platform.python_version()}")
-    bot.logger.info(f"Running on: {platform.system()} {platform.release()} ({os.name})")
-    bot.logger.info("-------------------")
-    basic_activity = discord.Activity(
-        name="you", type=discord.ActivityType.watching, status=discord.Status.online
-    )
-    await bot.change_presence(activity=basic_activity)
-    # TODO what the heck was this for
-    # if CONFIG.sync_commands_globally:
-    bot.logger.info("Syncing commands globally...")
-    # bot.tree.clear_commands(guild=bot.get_guild(CONFIG.server))
-    commands = await bot.tree.sync(guild=bot.get_guild(CONFIG.server))
-
-
-@bot.event
-async def on_message(message: discord.Message) -> None:
-    """
-    The code in this event is executed every time someone sends a message, with or without the prefix
-
-    :param message: The message that was sent.
-    """
-    if message.author == bot.user or message.author.bot:
-        return
-    await bot.process_commands(message)
-
-
-@bot.event
-async def on_command_completion(context: Context) -> None:
-    """
-    The code in this event is executed every time a normal command has been *successfully* executed.
-
-    :param context: The context of the command that has been executed.
-    """
-    full_command_name = context.command.qualified_name
-    split = full_command_name.split(" ")
-    executed_command = str(split[0])
-    if context.guild is not None:
-        bot.logger.info(
-            f"Executed {executed_command} command in {context.guild.name} (ID: {context.guild.id}) by {context.author} (ID: {context.author.id})"
-        )
-    else:
-        bot.logger.info(
-            f"Executed {executed_command} command by {context.author} (ID: {context.author.id}) in DMs"
-        )
-
-
-@bot.event
-async def on_command_error(context: Context, error) -> None:
-    """
-    The code in this event is executed every time a normal valid command catches an error.
-
-    :param context: The context of the normal command that failed executing.
-    :param error: The error that has been faced.
-    """
-    if isinstance(error, commands.CommandOnCooldown):
-        minutes, seconds = divmod(error.retry_after, 60)
-        hours, minutes = divmod(minutes, 60)
-        hours = hours % 24
-        embed = discord.Embed(
-            description=f"**Please slow down** - You can use this command again in {f'{round(hours)} hours' if round(hours) > 0 else ''} {f'{round(minutes)} minutes' if round(minutes) > 0 else ''} {f'{round(seconds)} seconds' if round(seconds) > 0 else ''}.",
-            color=0xE02B2B,
-        )
-        await context.send(embed=embed)
-    elif isinstance(error, exceptions.UserBlacklisted):
-        """
-        The code here will only execute if the error is an instance of 'UserBlacklisted', which can occur when using
-        the @checks.not_blacklisted() check in your command, or you can raise the error by yourself.
-        """
-        embed = discord.Embed(
-            description="You are blacklisted from using the bot!", color=0xE02B2B
-        )
-        await context.send(embed=embed)
-        bot.logger.warning(
-            f"{context.author} (ID: {context.author.id}) tried to execute a command in the guild {context.guild.name} (ID: {context.guild.id}), but the user is blacklisted from using the bot."
-        )
-    elif isinstance(error, exceptions.UserNotOwner):
-        """
-        Same as above, just for the @checks.is_owner() check.
-        """
-        embed = discord.Embed(
-            description="You are not the owner of the bot!", color=0xE02B2B
-        )
-        await context.send(embed=embed)
-        bot.logger.warning(
-            f"{context.author} (ID: {context.author.id}) tried to execute an owner only command in the guild {context.guild.name} (ID: {context.guild.id}), but the user is not an owner of the bot."
-        )
-    elif isinstance(error, commands.MissingPermissions):
-        embed = discord.Embed(
-            description="You are missing the permission(s) `"
-            + ", ".join(error.missing_permissions)
-            + "` to execute this command!",
-            color=0xE02B2B,
-        )
-        await context.send(embed=embed)
-    elif isinstance(error, commands.BotMissingPermissions):
-        embed = discord.Embed(
-            description="I am missing the permission(s) `"
-            + ", ".join(error.missing_permissions)
-            + "` to fully perform this command!",
-            color=0xE02B2B,
-        )
-        await context.send(embed=embed)
-    elif isinstance(error, commands.MissingRequiredArgument):
-        embed = discord.Embed(
-            title="Error!",
-            # We need to capitalize because the command arguments have no capital letter in the code.
-            description=str(error).capitalize(),
-            color=0xE02B2B,
-        )
-        await context.send(embed=embed)
-    else:
-        raise error
-
-
-async def load_cogs() -> None:
-    """
-    The code in this function is executed whenever the bot will start.
-    """
-    for file in (BASE_PATH / "cogs").iterdir():
-        if file.suffix == ".py":
-            extension = file.stem
-            try:
-                await bot.load_extension(f"cogs.{extension}")
-                bot.logger.info(f"Loaded extension '{extension}'")
-            except Exception as e:
-                exception = f"{type(e).__name__}: {e}"
-                bot.logger.error(f"Failed to load extension {extension}\n{exception}")
-
-
 asyncio.run(init_db())
-asyncio.run(load_cogs())
+asyncio.run(bot.load_cogs())
 
 bot.run(CONFIG.token)
